refactor: 자동차 경주 패키지 수정
