+++ src/main/java/racinggame/core/gasstation/GasStation.java
@@ -1,4 +1,4 @@
-package step3.racinggame.core.gasstation;
+package racinggame.core.gasstation;
 
 public interface GasStation {
